// app/src/bp-api/models/AccountConvertDustPayload.ts
/* generated using openapi-typescript-codegen -- do not edit */

/* istanbul ignore file */

/* tslint:disable */

/* eslint-disable */
import type { Asset } from './Asset'

export type AccountConvertDustPayload = {
  /**
   * The asset symbol to convert dust for.
   * If omitted, all dust balances will be converted.
   */
  symbol?: Asset
}