// app/src/bp-api/models/MaxOrderQuantity.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MaxOrderQuantity = {
  autoBorrow?: boolean
  autoBorrowRepay?: boolean
  autoLendRedeem?: boolean
  maxOrderQuantity: string
  price?: string
  side: string
  symbol: string
  reduceOnly?: boolean
}